2014.03.19 Ellenorzes openSUSE 13.1-en, flex/lemon prototypes. Az openSUSE 13.1-ben levo gcc 4.8.1 hianyolja az uj stilusu (parameterlistas) C deklaraciokat, deklaraciok potolva.

+++ tabobj/tbbtbtx/btsort.cpp
@@ -103,7 +103,7 @@ char *filemap(char *filename)
         stringnb(filename); _clp_taberrfilename(1); pop();
         _clp_taberror(0); pop();
 
-        fprintf(stderr,"filemap: mmap error %ld %d\n",mapLength,errno);
+        fprintf(stderr,"filemap: mmap error %ld %d\n",(long)mapLength,errno);
         exit(1);
     }
